pyproject.toml
[project]
name = "vibe-agency"
version = "1.0.0"
description = "File-based prompt framework for AI-assisted software project planning"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

authors = [
    { name = "vibe-agency team" }
]

# Dependencies from requirements.txt
dependencies = [
    "pyyaml>=6.0.1",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "google-api-python-client>=2.100.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
    "isort>=5.13.0",
    "yamllint>=1.35.0",
    "mypy>=1.7.0",
    "jsonschema>=4.17.0",
]

security = [
    "pip-audit>=2.6.0",
    "detect-secrets>=1.4.0",
]

hooks = [
    "pre-commit>=3.6.0",
]

# Install all extras with: uv sync --extra dev --extra security --extra hooks
all = [
    "vibe-agency[dev,security,hooks]",
]

[project.scripts]
vibe-cli = "vibe_cli:main"

# ============================================================================
# TOOL CONFIGURATIONS (Consolidated from .flake8, ruff.toml, etc.)
# ============================================================================

[tool.ruff]
line-length = 100
target-version = "py311"

# Ruff replaces: flake8, isort, black, bandit, etc.
[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "F",     # pyflakes
    "I",     # isort (import sorting)
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
    "UP",    # pyupgrade
    "S",     # flake8-bandit (security)
    "RUF",   # ruff-specific rules
    "C90",   # mccabe complexity
]

ignore = [
    "E501",  # line-too-long (handled by formatter)
    "S101",  # assert-used (OK in tests)
    "S314",  # xml parsing (we control the XML input in our use case)
    "S607",  # Starting a process with a partial executable path (git, gh, etc. are expected to be in PATH)
    "S110",  # try-except-pass (acceptable for optional features)
    "B904",  # raise-without-from-inside-except (acceptable for known error transformations)
    "RUF012",  # Mutable class attributes (acceptable for config dictionaries)
    "RUF013",  # implicit Optional (legacy code, can be fixed incrementally)
    "SIM102",  # Nested if statements (sometimes clearer than and)
    "SIM103",  # Return condition directly (sometimes if/else is clearer)
    "SIM105",  # contextlib.suppress (not always necessary for simple cases)
    "SIM115",  # Context manager for opening files (legacy patterns, can be fixed incrementally)
    "RUF001",  # Ambiguous unicode (emoji usage is intentional)
    "RUF005",  # List unpacking (legacy syntax is fine)
    "C901",   # Complex function (acceptable for main functions)
]

# Complexity checks (stricter than flake8's max-complexity=10)
[tool.ruff.lint.mccabe]
max-complexity = 15

# Per-file ignores
[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "E402",     # Allow late imports in tests
    "S101",     # Allow asserts in tests
    "S104",     # Allow binding to all interfaces in test configs
    "S108",     # Allow /tmp usage in tests
    "S603",     # Allow subprocess in tests
    "B007",     # Allow unused loop variables in tests
    "B011",     # Allow assert False in tests (for failure paths)
    "RUF043",   # Allow regex patterns without raw strings in tests
    "RUF059",   # Allow unused unpacked variables in tests
]
"test_*.py" = ["E402", "S101", "B007", "B011", "RUF043", "RUF059"]
"*-cli.py" = ["E402"]  # CLI scripts may need path manipulation
"tools/**/*.py" = ["B007"]  # Allow unused loop variables in tools
"scripts/**/*.py" = ["B007", "S603", "S108"]  # Allow utility patterns in scripts
"bin/**/*.py" = ["B007"]  # Allow unused loop variables in bin scripts
"docs/**/*.py" = ["S108"]  # Allow temp file usage in documentation examples

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ============================================================================
# PYTEST CONFIGURATION
# ============================================================================

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--disable-warnings",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

# ============================================================================
# MYPY CONFIGURATION (Type Checking)
# ============================================================================

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Start lenient, tighten gradually
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

# Gradually enable strict mode per module
[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

# ============================================================================
# BLACK CONFIGURATION (Code Formatting)
# ============================================================================

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # Defaults
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

# ============================================================================
# ISORT CONFIGURATION (Import Sorting)
# ============================================================================

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

# ============================================================================
# COVERAGE CONFIGURATION
# ============================================================================

[tool.coverage.run]
source = ["agency_os"]
omit = [
    "tests/*",
    "**/__pycache__/*",
    "**/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

# ============================================================================
# UV CONFIGURATION
# ============================================================================

[dependency-groups]
dev = []  # Managed via [project.optional-dependencies]

# ============================================================================
# BUILD SYSTEM
# ============================================================================

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agency_os", "lib"]
